Added 6-band measurements from 4-16-2015
Copied data from each band and processing code. The processing code is
using a different version of the refltool function library than is in
the "analysis" folder in this repository. We should rewrite the
processing code for this data to use the version of refltool in the
analysis folder, put that version of reltool into the lib folder, and
probably rewrite all future code to use that version.